+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- react-{identicon, qr}, ui-{keyring, settings}, vue-identicon 3.5.1
 - phishing 0.21.5
 - hw-ledger, keyring, util, util-crypto 12.3.1
 - extension-dapp 0.46.4